+++ lib/public/StormByte/logger/manipulators.cxx
@@ -22,6 +22,9 @@
 #include <StormByte/logger/implementation.hxx>
 #include <utility>
 namespace StormByte::Logger {
+	STORMBYTE_LOGGER_PUBLIC ComponentManip component(std::string name) {
+		return ComponentManip{std::move(name)};
+	}
 	STORMBYTE_LOGGER_PUBLIC GroupManip group(std::string name) {
 		return GroupManip{std::move(name)};
 	}